macosx: Minor fixes
Tickets: #1049, #1125 and #1164

+++ macosx/trunk/WebPlugIn/Sources/EstEIDReaderManager.mm
@@ -22,7 +22,7 @@ typedef struct _EstEIDReaderManagerState {
 		id <EstEIDReader> *data;
 		unsigned int count;
 	} readers;
-	unsigned int selection;
+	NSInteger selection;
 	NSTimer *timer;
 	id proxy;
 } EstEIDReaderManagerState;
